requirements.txt
# Bear AI Assignment - Essential Dependencies

# Web scraping (Stage 1)
selenium==4.15.2
webdriver-manager==4.0.1

# API framework (Stage 2) 
fastapi==0.104.1
uvicorn==0.24.0

# Database (Stage 2) - PostgreSQL as recommended in assignment
psycopg2-binary==2.9.7
sqlalchemy==2.0.23

# Environment variables
python-dotenv==1.0.0

# Testing
httpx==0.28.1 